requirements.txt
Flask>=3.0,<4
pywebview>=4,<5
